requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
azure-storage-blob==12.19.1
azure-core==1.30.1
dagster==1.7.11
dagit==1.7.11
